Merge the c_helper and f_helper into helper
A helper may have a C and Fortran part which required the helper to be
repeated in c_helper and f_helper. Now it only needs to be listed
once.

In addition, a c statements may be used to create the Fortran
interface for a C function. It defines the i_ fields for this
purpose.  When creating the non-bufferify wrapper it may still need a
f_helper. To avoid adding i_helper, just merge them together.

This shows up in strings.yaml when using the CFI wrapper. It did not
need the capsule, but the c_ wrapper still needs it to save memory
information for the std::string. It worked before because the
bufferify wrapper added it but with CFI it does not.

No changes to generate source. Only changes to JSON files.