Make std::string helper names specific to cdesc
A future commit will add CFI versions.  I though about having one true
version and instead of passing in a cdesc or CFI, pass in the parts
base_addr, elem_len, and size. But that made the wrapper more
complicated.  Part of the reason for a helper is to simplify the
wrapper.