Standardize Makefiles
Move some targets to defaults.mk.  Define variables and dependencies
in the Fortran Makefiles.  Define C_OBJS, F_OBJS and PY_OBJS for all
tests. It had slowly been migrating to that pattern.

Ensure that FFLAGS can be defined on the command line, while still
adding LOCAL_FFLAGS. Similar for CFLAGS, CXXFLAGS and LDFLAGS.

This will make it easier to add the ASAN flags to all tests.  The
flags is also needed on the load line.