Update PyYAML dependency for Python 3
The current limit on the PyYAML version does not work with newer
versions of Python. setup.py provides syntax to specify dependencies
based on python_version.

Also make a fix for Python 3.10 for collections.abc.Mapping.

Tested Python 2.7. It has rotted some and is not working.
Added a few fixes.

Found a recursive import: statements - whelpers - statements
Added a kludge by passing in statements module to the function in
whelpers that needed it.

Import fc-statements.py with pyYAML instead of json to avoid unicode
issues.

#392