Add cmemflags to the C/C++ capsule struct
The code is in wrapc.py but maybe it should be in fc-statements.json
as helpers. However, there is one capsule per class and not just one
for all so would need custom additional work.

Add options.F_assignment_api to meta[api] for ctor.  Needed to
initialize cmemflags.  This, also, feels kludgey.