Replace shared_ptr ctor and dtor
The Object's ctor and dtor may not exist or may have different
arguments. Remove them from the shared_ptr and create a new ctor and
dtor with no arguments.  May want to add arguments later. For
example. weak_ptr(shared_ptr).

Adding the ctor and dtor at the end of the list of functions reorders
the generated functions.

Rename has_ctor as allow_ctor in declast.py to make the usage clearer.