Set Fortran function owner attribute
Check conditions in set_func_owner_fortran and use
option.owner_default if appropriate.  To be used to set
capsule.cmemflags bit for SWIG_MEM_OWN.

By setting the owner metaattribute, the JSON files will change.  This
also changes the statement name which changes comments in the
generated code. pointer_owner is now the standard.

Add a default destructor for language=c.
Used with struct-c to delete struct-as-class constructor.