Add compiler/final directory
Contains a demonstration of the changes between gcc12 and gcc13 which
broke the shared.yaml test. It used the final procedure which is now
correctly called after an assignment.

From Modern Fortran explained incorporating Fortran 2023 page 318:
When a finalizable object is about to cease to exist [as] the variable
on the left-hand side of an intrinsic assignment statement.  The final
subroutine is invoked after the expression right-hand side has been
evaluated, but before it is assigned to the variable.