Fix another cxxlibrary.yaml createView fortran_generic test
createView is overloaded and uses a void* argument.  Test if it's
overloaded when deciding which wrappers to create.  Otherwise, two
wrappers may create identical name in the effort to minimize the
number of wrappers.

While this solves a problem, there may still be an issue with
cdesc.yaml GetScalar2 which may have a problem is its name was
overloaded since it uses a void pointer which is replaced in the
fortran_generic statement.