Change how option flatten_namespace is processed
The NamespaceNode is now always added to the parent NamespaceNode (or
LibraryNode).  Before it was not added if option.flatten_namespace was
set. Instead all of the namespaces declarations were added directly to
the parent Namespace. This could present issues with duplicate names.
Now the hierarchy is preserved in the Node tree.

This restores the splicer names which lost the namespace component in
some cases in an earlier commit. It also causes some reordering of
declarations in the C and Fortran wrappers.

Instead wrapf.py and wrapc.py are responsible for checking for
flatten_namespace and do the correct thing.  Note that Python and Lua
do not check for flatten_namespace and will create wrappers that
mirror the Node tree.

To be fixed: the 'Wrote filename' in the log is being duplicated since
write_file is being called too often due to now having to check for
flatten_namespace explicitly.