Changes to install to pypi
In HACKING.md, a 'Build a Python Distribution' before 'Create a Pull
Request for the Release'. Helps ensure that this distribution can
still build a pypi distribution.

For this release, work around a bit in twine by using a newer version
of packaging.

Add Makefile target 'do-test-install' to test a version of Shroud at
various parts of the release.

Include fewer files in MANIFEST.in. No need for test and such. If
needed then the user can download the github tar file.